move subgraph services to subgraph.service
