+++ src/components/controller/subgraph.ctrl.tsx
@@ -3,61 +3,17 @@ import { useAccount } from 'wagmi';
 import { useQueries, useQueryClient } from '@tanstack/react-query';
 import { Address } from 'viem';
 import config from '@/configuration';
-import { fetchSubgraph } from '@/services/subgraph.service';
-import { SubgraphQueryBuilder } from '@/lib/subgraph/subgraphQueryBuilder';
-import { transformSubgraphData } from '@/lib/subgraph/subgraphDataTransform';
-import { ISubgraphState } from '@/types/subgraph';
-
-const SUBGRAPH_METADATA_QUERY = `{
-  _meta {
-	block {
-	  number
-	}
-  }
-}`;
-
-export const fetchSubgraphData = async (
-	chainId?: number,
-	address?: Address,
-): Promise<ISubgraphState> => {
-	if (!chainId || !address) return {} as ISubgraphState;
-	let response;
-	let uri = config.EVM_NETWORKS_CONFIG[chainId]?.subgraphAddress;
-
-	if (!uri) {
-		response = {};
-	} else {
-		response = await fetchSubgraph(
-			SubgraphQueryBuilder.getChainQuery(chainId, address),
-			chainId,
-		);
-	}
-	return transformSubgraphData({
-		...response,
-		networkNumber: chainId,
-	});
-};
-
-export const fetchLatestIndexedBlock = async (chainId: number) => {
-	let response;
-	let uri = config.EVM_NETWORKS_CONFIG[chainId]?.subgraphAddress;
-
-	if (!uri) {
-		response = {};
-	} else {
-		response = await fetchSubgraph(SUBGRAPH_METADATA_QUERY, chainId);
-		return response._meta.block.number;
-	}
-	return response;
-};
+import {
+	fetchLatestIndexedBlock,
+	fetchSubgraphData,
+} from '@/services/subgraph.service';
 
 const SubgraphController: React.FC = () => {
-	const { chain, address } = useAccount();
-	const chainId = chain?.id;
+	const { address } = useAccount();
 	const queryClient = useQueryClient();
 	const pollingIntervalsRef = useRef<{ [key: string]: NodeJS.Timeout }>({});
 
-	const queries = useQueries({
+	useQueries({
 		queries: config.CHAINS_WITH_SUBGRAPH.map(chain => ({
 			queryKey: ['subgraph', chain.id, address] as [
 				string,
