+++ src/lib/subgraph/subgraph.queries.ts
@@ -1,0 +1,7 @@
+export const SUBGRAPH_METADATA_QUERY = `{
+	_meta {
+	  block {
+		number
+	  }
+	}
+  }`;
